Changed the type of 'id' from Integer to Long

+++ src/main/java/com/thedevhorse/cleanarchitecture/infra/repository/AthleteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AthleteRepository extends JpaRepository<AthleteEntity, Integer> {
+public interface AthleteRepository extends JpaRepository<AthleteEntity, Long> {
 
     Optional<AthleteEntity> findByAthleteId(String athleteId);
 }
